Fix circle ci build issues
We are getting errors on circle ci builds (e.g. see [1]):

```
CircleCI was unable to run the job runner because we were unable to
execute commands in build container.

This typically means that the build container entrypoint or command is
terminating the container prematurely, or interfering with executed
commands.  Consider clearing entrypoint/command values and try again.
```

Folks in this thread [2] suggest removing the `command: /sbin/init` line
initially added by the v1 => v2 migration script.

---

1: https://circleci.com/gh/zsh-users/zsh-autosuggestions/381
2: https://discuss.circleci.com/t/circleci-was-unable-to-run-the-job-runner/31894/18