Fix issues with LASTWIDGET caused by async suggestions
There are a number of widgets that rely on LASTWIDGET so that they can
chain their behavior together on subsequent invocations. When a
suggestion was fetched asynchronously, the widget showing the suggestion
would overwrite LASTWIDGET and break these widgets that rely on it.
Relatively recently, a flag was added to the `zle` builtin to opt out of
setting LASTWIDGET when calling a widget. We can use that flag to avoid
setting LASTWIDGET when displaying suggestions that were fetched
asynchronously.